requirements.txt
tesseract
ghostscript
pngquant
jbig2enc
qrcode
PyPDF2
PdfReader
pillow
yagmail
python-dotenv
openai-whisper
ffmpeg-python
llama-cpp-python
icalendar
transformers
sentencepiece
sacremoses
pydub
pdf2docx
moviepy-editor